feat(ios): Add `hasTargetFrame` prop on `onShouldStartLoadWithRequest`
On iOS, current implementation of `onShouldStartLoadWithRequest` does
not allow to differentiate links that target the current WebView from
ones that target a new WebView (i.e. `target="_blank"` links).

In some scenario we want to have this info to prevent the WebView
navigation

This info can be useful if `onOpenWindow` event is not registered.
Otherwise this event will be called instead of
`onShouldStartLoadWithRequest`