feat(android): Add `onOpenWindow` event
On Android, default behavior is to open the device's browser when
`window.open('http://someurl, '_blank')` is called from JS or when a
link with `target="_blank"` is clicked

In some scenario we want to be able to intercept this from react-native
and to handle the link by ourselves (i.e. to open it into an InApp
Browser)

The `onOpenWindow` event is now called on those scenario. In this case,
the default "open browser" scenario does not occur