feat(iOS): Add `onOpenWindow` event
On iOS, default behavior is to navigate the WebView to the target URL
when `window.open('http://someurl, '_blank')` is called from JS

In some scenario we want to be able to intercept this from react-native
and to handle the link by ourselves (i.e. to open it into an InApp
Browser)

The `onOpenWindow` event is now called in this scenario. Then the
default "open browser" scenario does not occur